requirements.txt
fastapi==0.75.0
uvicorn==0.16.0
pydantic==1.9.0
openai==1.23.6
requests==2.27.1
python-dotenv==0.19.1
